***CODE COMPILES ONCE AGAIN.
The main offender was the fact that in MainLoopC, we were using the wrong type for the Counter.  By changing it from uint32_t to uint16_t the code compiled for the real Autopilot.  Also, fixes for the Simulation are committed here.  Both the simulation and actual autopilot application should compile now, without errors.